move redis driver test to own file
